requirements2.txt
# without transform2.py

faiss-cpu==1.8.0
fastapi==0.111.0
InstructorEmbedding==1.0.1
langchain==0.1.20
langchain_community==0.0.38
langchain_core==0.1.52
python-dotenv==1.0.1
sentence-transformers==2.2.2
starlette==0.37.2
streamlit==1.34.0
streamlit_datalist==0.0.5
